requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cron==0.6.0
django-crontab==0.7.1
django-filter==22.1
django-otp==1.1.6
djangorestframework==3.14.0
idna==3.4
mailgun==0.1.1
pyactiveresource==1.0.1
pycodestyle==2.10.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
